- В авторежиме с перехватом хотспота при нажатии «Запустить» приложение теперь сразу проверяет, запущено ли оно от имени администратора. Если прав нет, вместо непонятной ошибки появляется окно с кнопкой «Перезапустить от администратора» — нажмите её, подтвердите запрос UAC, и VoxelProxy перезапустится с нужными правами. - Отключено стандартное контекстное меню браузера (правый клик с пунктами «Назад», «Обновить», «Сохранить как», «Печать» и т.д.), которое не относится к приложению. В текстовых полях правый клик с копированием/вставкой по-прежнему работает.

+++ src-tauri/src/events.rs
@@ -42,6 +42,6 @@ pub struct ServerAddrEvent(pub String);
 
 #[derive(Serialize, Clone, Type, Event)]
 pub struct UpdateProgressEvent {
-    pub downloaded: u64,
-    pub total: u64,
+    pub downloaded: u32,
+    pub total: u32,
 }
